refactor: move get user to new router
