+++ src/action/auth.ts
@@ -1,8 +1,7 @@
 "use server";
 
-import { auth } from "../server/api/auth";
 import { cookies } from "next/headers";
-import { serverapi } from "@/trpc/server";
+import { auth } from "../server/api/auth";
 
 export async function signIn(formData: FormData) {
   try {
@@ -26,15 +25,5 @@ export async function signIn(formData: FormData) {
   } catch (error) {
     console.error(error);
     return "Invalid email or password";
-  }
-}
-
-export async function GetAllParticipant() {
-  try {
-    const data = await serverapi.auth.getParticipants.query();
-    return data;
-  } catch (err) {
-    console.error(err);
-    throw err;
   }
 }
