+++ src/server/api/root.ts
@@ -2,6 +2,7 @@ import { createTRPCRouter } from "@/server/api/trpc";
 import { authRouter } from "./routers/auth";
 import { chatRouter } from "./routers/chat";
 import { profileRouter } from "./routers/profile";
+import { userRouter } from "./routers/user";
 
 /**
  * This is the primary router for your server.
@@ -12,6 +13,7 @@ export const appRouter = createTRPCRouter({
   chat: chatRouter,
   auth: authRouter,
   profile: profileRouter,
+  user: userRouter,
 });
 
 // export type definition of API
